# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "College Hierarchy API",
    "version": "1.0.0",
    "description": "API documentation for the College Hierarchy system"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local Server"
    }
  ],
  "paths": {
    "/principal/addprincipal": {
      "post": {
        "summary": "Add a new Principal",
        "description": "Creates a new principal in the college hierarchy.",
        "operationId": "addPrincipal",
        "tags": ["Principal"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Dr. Krishna"
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Principal created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Principal added successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer", "example": 1 },
                        "name": { "type": "string", "example": "Dr. Krishna" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Name is not Provided" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": false },
                    "message": { "type": "string", "example": "Error in adding principal" },
                    "error": { "type": "string", "example": "Database connection failed" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/principal/getprincipal": {
      "get": {
        "summary": "Get Principal",
        "description": "Retrieves principal data.",
        "operationId": "getPrincipal",
        "tags": ["Principal"],
        "responses": {
          "200": {
            "description": "Principal fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Principal fetched successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer", "example": 1 },
                        "name": { "type": "string", "example": "Dr. Krishna" }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Error in getting principal" },
                    "error": { "type": "string", "example": "Database connection failed" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/principal/getbyid/{id}": {
      "get": {
        "summary": "Get Principal by ID",
        "description": "Retrieves principal data by ID.",
        "operationId": "getPrincipalById",
        "tags": ["Principal"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Principal fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Principal with id: 1 fetched successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer", "example": 1 },
                        "name": { "type": "string", "example": "Dr. Krishna" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Invalid Id format" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Principal not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Principal with id not found" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Error in getting principal" },
                    "error": { "type": "string", "example": "Database connection failed" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/principal/deleteprincipal/{id}": {
      "delete": {
        "summary": "Delete Principal",
        "description": "Deletes a principal and all related child data.",
        "operationId": "deletePrincipal",
        "tags": ["Principal"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Principal and child data deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Principal and child data deleted successfully" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Id is required" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Principal not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Id not found or delete failed" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Error while deleting principal" },
                    "error": { "type": "string", "example": "Database error" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}